Added some quests to the Database.json. Created a QuestDataFactory class: which creates Quests and the other objects it composes (Rewards, Tasks). Created a QuestManager class: stores all the quests created using the factory with the information from the Json file. Also stores all the methods and actions required by the Quest System for the game (it will be last layer interacting directly with the game and the player). Some methods still remain to be completed to connect the Quest System with the rest of the game. Corrected the other classes and took into account most of the comments made.
Explanation on how the Quest System is going to work (aim):
Player inputs "quests", which will result in displaying a numbered list of the quests that exists.
Player should then be able to enter a number (number of the quest in the list that he would like to complete).
The program will then attempt to complete quest and update the whole list of quests again (displaying it again as it did when the player entered "quests").

What is left to be done by wednesday 30rd November:
Create a new reward type: inventory, however, have not seen how it works, I could not write this part.
Finish to connect the Quest System with the rest of the game
Need to finish polishing the code for the
Writing down the tests where needed for the Quest System.