Created the basis for the interaction with the database for the quest.
QuestDatabase.json contains the information for the quest (in Json format).
QuestData contains all the information required to create a quest.
QuestDataFactory creates a QuestData object from a Json file.
QuestDataManager looks for the through the QuestDatabase.json to fetch the Quest we are looking through and returns a QuestData object.

Integrated the Quests to the event system.

Remains to be updated: in EncounterEvent, add the quest to the quests undertaken by the player.